Remove duplicate test file
`test_codegen_c_host_fadd.py` accidentally made it into the last
MicroTVM PR (and stole a test from `test_codegen_c_host.py`???).
